feat : 팀 아이디 자료형 변경

+++ backend/member_service/src/main/java/yare/yare/domain/member/dto/response/MemberDetailsRes.java
@@ -21,7 +21,7 @@ public class MemberDetailsRes {
 
     private String email;
 
-    private Long myTeamId;
+    private Integer myTeamId;
 
     private String myTeamName;
